Improved authorization and registration Separated controllers: LogController, UserAccountController, DiagramController Added disappearing messages Fixed the markup

+++ src/main/java/com/example/project/service/UserService.java
@@ -2,7 +2,6 @@ package com.example.project.service;
 
 import com.example.project.model.Dto.DiagramDto;
 import com.example.project.model.Entity.DesignTheme;
-import com.example.project.model.Dto.UserDto;
 import com.example.project.model.Entity.Diagram;
 import com.example.project.model.Entity.User;
 import com.example.project.repository.UserRepository;
@@ -11,7 +10,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
 import java.util.Optional;
 
 
@@ -24,34 +22,42 @@ public class UserService {
 
 
     @Autowired
-    public UserService(UserRepository userRepository, DesignThemeService designThemeService, DiagramService diagramService) {
+    public UserService(UserRepository userRepository, DesignThemeService designThemeService,
+                       DiagramService diagramService) {
         this.userRepository = userRepository;
         this.designThemeService = designThemeService;
         this.diagramService = diagramService;
+    }
+
+    public Optional<User> getById(Long id) {
+        return userRepository.findById(id);
+    }
+
+    public String getEmailById(Long id) {
+        if (getById(id).isPresent()) {
+            User user = getById(id).get();
+            return user.getEmail();
+        }
+        return null;
     }
 
     public Optional<User> getByEmail(String email) {
         return userRepository.findByEmail(email);
     }
 
-    public void createUser(UserDto userDto) {
-        DesignTheme designTheme = designThemeService.getByName(userDto.getDesignTheme());
-
+    public void createUser(String email, String password, String designThemeName) {
+        DesignTheme designTheme = designThemeService.getByName(designThemeName);
         User user = new User();
-        user.setEmail(userDto.getEmail());
-        user.setPassword(userDto.getPassword());
+        user.setEmail(email);
+//        user.setPassword(bCryptPasswordEncoder.encode(password));
+        user.setPassword(password);
         user.setDesignTheme(designTheme);
-
         userRepository.save(user);
     }
 
-    public Long addDiagram(String email, DiagramDto diagramDto) {
-        if (getByEmail(email).isPresent()) {
-            User user = getByEmail(email).get();
-            Diagram diagram = diagramService.createByName(diagramDto, user);
-            return diagram.getId();
-        }
-        return null;
+    public Long addDiagram(User user, DiagramDto diagramDto) {
+        Diagram diagram = diagramService.createByName(user, diagramDto);
+        return diagram.getId();
     }
 
     public void changePassword(String email, String password) {
@@ -62,21 +68,25 @@ public class UserService {
         }
     }
 
-    public void changeDesignTheme(String email, String designTheme) {
-        if (getByEmail(email).isPresent()) {
-            User user = getByEmail(email).get();
+    public void changeDesignTheme(User user, String designTheme) {
+        if (designTheme != null) {
             user.setDesignTheme(designThemeService.getByName(designTheme));
             userRepository.save(user);
         }
     }
 
-    public void deleteDiagram(Long diagramId) {
-        diagramService.deleteDiagram(diagramId);
+    public void deleteDiagram(Long userId, Long diagramId) {
+        //Проверка является ли отправитель запроса владельцем диаграммы
+        if(diagramService.getByID(diagramId).getUserId().equals(userId)){
+            diagramService.deleteDiagram(diagramId);
+        }
     }
 
     @Transactional
-    public void deleteUser(User user) {
-        userRepository.deleteAllDiagramByUser(user.getId());
-        userRepository.delete(user);
+    public void deleteUser(Long userId) {
+        if(getById(userId).isPresent()) {
+            userRepository.deleteAllDiagramByUser(userId);
+            userRepository.delete(getById(userId).get());
+        }
     }
 }
